requirements.txt
# API Framework
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1

# ML & NLP
torch==2.2.0
transformers==4.37.2
pyod==1.1.3
shap==0.44.0
langchain==0.1.9
openai==1.12.0
whisper==1.1.10

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 